requirements.txt
Flask
Flask_Session
transformers
torch
python-dotenv
werkzeug